% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_genes_by_donors.R
\name{arrange_genes_by_donors}
\alias{arrange_genes_by_donors}
\title{Title}
\usage{
arrange_genes_by_donors(count_matrix, meta_ind, meta_cell)
}
\arguments{
\item{count_matrix}{}

\item{meta_ind}{}

\item{meta_cell}{}
}
\description{
Title
}
