% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_array_list.R
\name{dist_array_list}
\alias{dist_array_list}
\title{To initialize a list of arrays, dist_array_list}
\usage{
dist_array_list(dat_res)
}
\arguments{
\item{dat_res}{}
}
\description{
To initialize a list of arrays, dist_array_list
}
