% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_dist_array.R
\name{.form_dist_array}
\alias{.form_dist_array}
\title{Form a 3D Distance Array from a List of Distance Matrices}
\usage{
.form_dist_array(dist_list)
}
\arguments{
\item{dist_array_list}{A named list of distance matrices. Each element of the list
corresponds to a gene and contains a 3D array where the third dimension includes
the \code{"was2"} distance.}
}
\value{
A 3D array with dimensions \code{(n_genes, n_donors, n_donors)},
where \code{n_genes} is the number of genes and \code{n_donors} is the number
of donors. The array is indexed by gene and contains the \code{"was2"} distances
between donors.
}
\description{
This internal function constructs a 3D array from a list of distance matrices,
where each matrix corresponds to a specific gene and contains pairwise distances
between donors.
}
\examples{
# Example usage (assuming `dist_list` is a valid input)
dist_array <- .form_dist_array(dist_list)

}
\keyword{internal}
