% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housekeeping_lin_df}
\alias{housekeeping_lin_df}
\title{Lin et al. housekeeping genes}
\format{
\subsection{\code{housekeeping_lin}}{

A data frame with 12986 rows (genes) and 3 columns:
\describe{
\item{Gene}{Gene name}
\item{Stability.index}{Stability index, the calculation the authors made to derive if a gene is "stable" or not}
\item{Housekeeping}{Internal criteria defined as Stability.index larger that 0.8}
}
}
}
\source{
\url{https://academic.oup.com/gigascience/article/8/9/giz106/5570567}
}
\usage{
housekeeping_lin_df
}
\description{
From https://academic.oup.com/gigascience/article/8/9/giz106/5570567,
giz106_supplemental_table.xlsx.
}
\details{
Using the criterion that \code{Stability.index} should be above 0.8, there are 759 housekeeping genes
}
\examples{
\dontrun{
# How the data was loaded
df <- openxlsx::read.xlsx("giz106_supplemental_table.xlsx",
sheet = "human development")
colnames(df)[1] <- "Gene"
df$Housekeeping <- (df$Stability.index > 0.8)
df <- df[,c("Gene", "Stability.index", "Housekeeping")]
housekeeping_lin_df <- df
usethis::use_data(housekeeping_lin_df)

# To load
housekeeping_lin <- housekeeping_lin_df$Gene[housekeeping_lin_df$Housekeeping]
}
}
\keyword{datasets}
