% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_dist_custom.R
\name{ideas_dist_custom}
\alias{ideas_dist_custom}
\title{Using the divergence function many many many times,
and putting the correct values into the empty
dist_array_list (list of arrays)
and then reorganize it to result_array_list}
\usage{
ideas_dist_custom(
  count_input,
  meta_cell,
  meta_ind,
  var_per_cell,
  var2test,
  var2test_type = c("binary", "continuous"),
  per_cell_adjust = c("NB", "both")
)
}
\arguments{
\item{count_input}{}

\item{meta_cell}{}

\item{meta_ind}{}

\item{var_per_cell}{}

\item{var2test}{}

\item{var2test_type}{}

\item{per_cell_adjust}{}
}
\description{
Using the divergence function many many many times,
and putting the correct values into the empty
dist_array_list (list of arrays)
and then reorganize it to result_array_list
}
