% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microglia_prater_df}
\alias{microglia_prater_df}
\title{Sun et al. microglia AD genes}
\format{
\subsection{\code{microglia_prater_df}}{

A data frame with 942 rows (genes) and 9 columns:
\describe{
\item{Gene}{Gene name}
\item{baseMean}{Average mean across all donors}
\item{log2FoldChange}{Log2 fold change between AD and non-AD}
\item{lfcSE}{Standard error of the log fold change}
\item{stat}{Test statistic from MAST}
\item{pvalue}{P-value from MAST}
\item{padj}{Adjust p-value from MAST}
\item{cluster}{The microglia cluster that the DEG analysis was performed on}
\item{DEG}{Internal criteria to deem a gene as DEG when looking at microglia as a whole}
}
}
}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0092867423009716}
}
\usage{
microglia_prater_df
}
\description{
From https://www.nature.com/articles/s43587-023-00424-y,
43587_2023_424_MOESM3_ESM.xlsx.
}
\details{
Using the criterion that a gene is deemed DEG if its adjusted p-value is smaller than 0.05 in any microglia cluster
}
\examples{
\dontrun{
# How the data was loaded
df <- openxlsx::read.xlsx("43587_2023_424_MOESM3_ESM.xlsx",
sheet = "AD_vs_Ctrl_DEGs",
startRow = 3)
microglia_prater_df <- df
enriched_genes <- unique(microglia_prater_df$Gene[microglia_prater_df$padj <= 0.05])
microglia_prater_df$DEG <- (microglia_prater_df$Gene \%in\% enriched_genes)
usethis::use_data(microglia_prater_df)

# To load
microglia_prater <- unique(microglia_prater_df$Gene[microglia_prater_df$DEG])
}
}
\keyword{datasets}
