% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_array_list.R
\name{result_array_list}
\alias{result_array_list}
\title{Generate Result Array List}
\usage{
result_array_list(dist_array_list, meta_ind)
}
\arguments{
\item{dist_array_list}{}

\item{meta_ind}{}
}
\value{
A list of 3-dimensional arrays, one for each metric (distance, location, location_sign, size, size_sign, shape).
Each array has dimensions (number of genes, number of donors, number of donors) and contains the respective metric values.
}
\description{
Generate Result Array List
}
