% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/was2code_lfc.R
\name{was2code_lfc}
\alias{was2code_lfc}
\title{Compute standardized logFC-style effect sizes from Was2CoDE output}
\usage{
was2code_lfc(dist_list, case_idx, control_idx)
}
\arguments{
\item{dist_list}{A list of distance arrays (from was2code_dist)}

\item{case_idx}{Integer vector of row indices corresponding to cases}

\item{control_idx}{Integer vector of row indices corresponding to controls}
}
\value{
A matrix with genes as rows and distance components as columns:
columns = c("was2", "location", "size", "shape")
}
\description{
Compute standardized logFC-style effect sizes from Was2CoDE output
}
