% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/was2de_pvalue.R
\name{was2de_pvalue}
\alias{was2de_pvalue}
\title{Perform Wilcoxon tests on metrics for Var groups}
\usage{
was2de_pvalue(dist_list, meta_ind, Var, verbose = 0)
}
\arguments{
\item{dist_list}{List of distance matrices}

\item{meta_ind}{Data frame containing metadata information for individuals}

\item{Var}{dependent variable used for wilcoxon test}
}
\value{
A matrix with rows representing genes and columns representing means of distance metrics for different groups and p-values
}
\description{
Perform Wilcoxon tests on metrics for Var groups
}
