% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence.R
\name{divergence}
\alias{divergence}
\title{define divergence() using wasserstein1d()}
\usage{
divergence(a, b, p = 2)
}
\arguments{
\item{a}{Numeric (gene expressions) vector.}

\item{b}{Numeric (gene expressions) vector.}

\item{p}{Integer, power parameter for the Wasserstein distance.}
}
\description{
define divergence() using wasserstein1d()
}
