% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideas_dist_custom.R
\name{ideas_dist_custom}
\alias{ideas_dist_custom}
\title{Compute Wasserstein-2 decomposition for Gene Expression Data
Using the divergence function many many many times,
and putting the correct values into the empty
dist_array_list (list of arrays)
and then reorganize it to result_array_list}
\usage{
ideas_dist_custom(
  count_input,
  meta_cell,
  meta_ind,
  var_per_cell,
  var2test,
  var2test_type = c("binary", "continuous"),
  verbose = 0
)
}
\arguments{
\item{count_input}{}

\item{meta_cell}{}

\item{meta_ind}{}

\item{var_per_cell}{}

\item{var2test}{}

\item{var2test_type}{}
}
\value{
A list of 3-dimensional arrays, each corresponding to a distance metric (distance, location, location_sign, size, size_sign, shape).
Each array has dimensions (number of genes, number of donors, number of donors) and contains the respective metric values.
}
\description{
Compute Wasserstein-2 decomposition for Gene Expression Data
Using the divergence function many many many times,
and putting the correct values into the empty
dist_array_list (list of arrays)
and then reorganize it to result_array_list
}
